requirements.txt
requests==2.32.3
hubspot-api-client
python-dotenv
